requirements.txt
py-trees==0.8.3
networkx==2.2
Shapely==2.0.6
psutil
xmlschema==1.0.18
ephem
tabulate
opencv-python==4.2.0.32
numpy
matplotlib
six
simple-watchdog-timer
lxml
sympy
pyyaml
scipy
lmdb
omegaconf
munch
openai==0.28
